Draft of new layout for pypowsybl.loadflow module
