+++ pypowsybl/loadflow/__init__.py
@@ -1,0 +1,21 @@
+# Brings relevant types and methods into public namespace
+from .impl.loadflow import (
+    run_ac,
+    run_dc,
+    ComponentStatus,
+    ComponentResult,
+    Parameters,
+    ConnectedComponentMode,
+    BalanceType,
+    VoltageInitMode,
+    get_provider_parameters_names,
+    get_default_provider,
+    get_provider_names,
+    set_default_provider
+)
+
+from .impl.validation import (
+    run_validation,
+    ValidationType,
+    ValidationResult,
+)
